Rename README.md to Description.md
This project helps us create an automated attendence system using Opencv and Python. 
The image of the student is stored in the database along with their student ID number. During the time of attendence the student inputs their Student ID number a webcam window is opened and the image of the student is captured .This image is resized and converted into gray-scale and compared with the image in the database.